Agregar casteo de constante `EXTREMO_INFINITO` a `size_t`

+++ Utils/ValidadoresJSON/Condiciones/Indices/CLASE_CondicionRangoIndices.h
@@ -7,7 +7,7 @@
 #ifndef CONDICION_RANGO_INDICES
 #define CONDICION_RANGO_INDICES
 
-#define EXTREMO_INFINITO (-1)
+#define EXTREMO_INFINITO ((size_t) -1)
 
 #include "../CLASE_CondicionValidador.h"
 #include "../../../FuncionesGlobales.h"
